move utility to javascript-utilities file
