+++ src/services/javascript-utilities.js
@@ -76,11 +76,34 @@ const splitArray = (array, testFunc) => {
   return [left, right];
 };
 
+/**
+ * Checks whether an object is instance of a given class
+ * @param {Object} instance  the instance object to check
+ * @param klass     the class the instance object to be checked against
+ * @returns {boolean} whether the instance is instanceof the provided klass
+ */
+const instanceOf = (instance, klass) => {
+  while (instance && instance !== Object.prototype) {
+    if (!instance || !instance.constructor || !instance.constructor.name) {
+      return false;
+    }
+
+    if (klass.name === instance.constructor.name) {
+      return true;
+    }
+
+    instance = Object.getPrototypeOf(instance);
+  }
+
+  return false;
+};
+
 module.exports = {
   doesObjectHaveProperty,
   translateObject,
   translateKeys,
   translateValues,
   sleep,
-  splitArray
+  splitArray,
+  instanceOf,
 };
